Move generators into their own dirs
