+++ codegeneration/python/Generator.js
@@ -1,12 +1,12 @@
 const path = require('path');
-const CodeGenerator = require('./CodeGenerator.js');
+const CodeGenerator = require('../javascript/Visitor.js');
 
 const {
   SemanticArgumentCountMismatchError,
   SemanticGenericError,
   SemanticTypeError
 } = require(path.resolve('helper', 'error'));
-const {Types} = require('./SymbolTable');
+const {Types} = require('../SymbolTable');
 
 /**
  * This Visitor walks the tree generated by parsers and produces Python code.
